fixed bugs & integrated trip conversation

+++ frontend/lib/services/rating_service.dart
@@ -1,13 +1,12 @@
 import 'package:carpool/models/user.dart';
 import 'package:carpool/services/api_service.dart';
-import 'package:dio/dio.dart';
 
 class RatingService {
   RatingService._();
 
   static Future<int> getRating(String id) async {
     final response = await ApiService.instance.get('/ratings/users/$id');
-    return response.data as int;
+    return (response.data as double).toInt();
   }
 
   static Future<User> getUser(String id) async {
